refactor(Tabs): remove the interface to index file
