--- src/components/navigation/Tabs/interfaces/Tabs.Item.interface.ts
@@ -1,5 +1,0 @@
-export interface ITabsItem {
-  id: string;
-  label: string;
-  isDisabled: boolean;
-}
