+++ src/components/navigation/Tabs/stories/TabsResponsiveController.tsx
@@ -1,7 +1,6 @@
 import { useState, useEffect } from "react";
-import { Tabs } from "../index";
+import { Tabs, ITabsProps } from "..";
 import { Types, types } from "../types/Tabs.type";
-import { ITabsProps } from "../interfaces/Tabs.interface";
 
 const TabsResponsiveController = (props: ITabsProps) => {
   const { tabs, selectedTab } = props;
